# swagger.json
{
  "info": {
    "title": "paris-js-0918-groupe7-node",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "Api to manage questions / answers"
  },
...
  "paths": {
...
    "/companies/{id(\\d+)}": {
      "get": {
        "tags" : ["companies"],
        "summary": "/companies/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "tags" : ["companies"],
        "summary": "/companies/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "tags" : ["companies"],
        "summary": "/companies/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/companies/{name}": {
      "get": {
        "tags" : ["companies"],
        "summary": "/companies/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ]
      }
    },
...
    "/users/{id(\\d+)}": {
      "get": {
        "tags" : ["users"],
        "summary": "/users/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "tags" : ["users"],
        "summary": "/users/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "tags" : ["users"],
        "summary": "/users/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/users/{email}": {
      "get": {
        "tags" : ["users"],
        "summary": "/users/{email}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true
          }
        ]
      }
    },
...
    "/answers_possibilities/{id(\\d+)}": {
      "get": {
        "tags" : ["answers_possibilities"],
        "summary": "/answers_possibilities/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "tags" : ["answers_possibilities"],
        "summary": "/answers_possibilities/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "tags" : ["answers_possibilities"],
        "summary": "/answers_possibilities/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      }
    },
...
    "/questions/{id(\\d+)}": {
      "get": {
        "tags" : ["questions"],
        "summary": "/questions/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "tags" : ["questions"],
        "summary": "/questions/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "tags" : ["questions"],
        "summary": "/questions/{id(\\d+)}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id(\\d+)",
            "in": "path",
            "required": true
          }
        ]
      }
    },
...
  },
  "swagger": "2.0"
}